Complete Milestone 1: Projects 1 - 3 (#1)
* Add Milestone1 starter files

* Add Flag assets under Bundle

This is done through creating new Group under Milestone1, and then
dragging all the image assets under this Group

* Initialise TableView with Flag file names

Note: Don't forget to change the identifier in the Storyboard for the
TableView cell!

* Setup Flag class to do capitalisation mapping

Perhaps hardcoding might not be the way to go for the hashmap. An
alternative could be using a capitalise() method, but then adding
special cases for like US and UK?

* Add NavigationController together with the DetailViewController to show the flag

* Modularise Flag methods to get the trimmed flag name

* Touch up Navigation Bar titles

* Add borders and resize flag to show borders

Need to figure out a way to adjust the size of the ImageView
dynamically. Currently there are trailing whitespaces.

* Add sharing photos feature

* Fix Auto Layout constraints of the Flags

The trick to obtain an ImageView that fits nicely is to:
1. Align Horizontally and Vertically in Container
2. Add new constraints. The important step is to tick 'Aspect Ratio'
3. Set 0 Margin for both Leading and Trailing