Challenge 3: Start with a TableView before using WebKit
A lesson for me is perhaps to remove the loadView, because this method
helps to setup the webview to the view. However, we want to delay
switching the view to webview only after the table cell is clicked.

The rest is about moving the functionalities to the appropriate places
in the other methods.