Bonus: Fix bugs pertaining to word cases
This is the laziest way to fix the problem, because it will force
whatever input to become lowercase.

An alternative way to do this is to really put the answer the way user
inputs it, but in the `isOriginal` method, you need to create a copy of
the array with all the words lowercased. In terms of time complexity,
this is not really efficient. I suppose this is a trade-off that can be
made.