Bonus: Load list of websites from a file
An interesting "bug" cropped up: Every file written in Xcode ends
with a newline character. This is a problem because I separate the components in the read string by \n. As a result, I added an additional check to remove empty strings at the end of the array right after it was read.

I could have just created a clean start.txt file, but I think many text
editors these days would have the newline characters at the end of the
file, and it is best to handle this possibility inside the code rather
than forcing the files to conform to our needs.