Complete Project 4 (#2)
* Add Project4 starter files

* Finish basic Project4

* Challenge 1: Show alert if visiting blocked website

I'm not quite sure how to handle the about:blank, which is present in
Apple's website. For now I have allowed that to be loaded, because there
is nothing wrong about it I think.

* Challenge 2: Add Back and Forward buttons

I couldn't find proper SystemItem for back and forward buttons, so I
suppose I should just use plain text for this assignment. One thing
which tripped me up a little bit was I used `self` for target rather
than `webView` for these buttons. I suppose `webView` makes sense
because I am only moving between the WebView pages.

* Challenge 3: Start with a TableView before using WebKit

A lesson for me is perhaps to remove the loadView, because this method
helps to setup the webview to the view. However, we want to delay
switching the view to webview only after the table cell is clicked.

The rest is about moving the functionalities to the appropriate places
in the other methods.

* Bonus: Load list of websites from a file

An interesting "bug" cropped up: Every file written in Xcode ends
with a newline character. This is a problem because I separate the components in the read string by \n. As a result, I added an additional check to remove empty strings at the end of the array right after it was read.

I could have just created a clean start.txt file, but I think many text
editors these days would have the newline characters at the end of the
file, and it is best to handle this possibility inside the code rather
than forcing the files to conform to our needs.