Complete Project 5 (#3)
* Add Project5 starter files

* Complete basic Project5

* Challenge 1: Disallow shorter than 3 letters and using same word

The choice of methods to place these checks under might not be optimal,
but I suppose this is the only way to prevent more if-else checks

* Challenge 2: Refactor method to show error messages

Another option is to just call the `showErrorMessage` directly in each
of the `else` blocks, but I think what we have is nicer as it allows the
Strings to be more easily modified. Putting the String in
showErrorMessage directly might make it inconvenient to change.

* Challenge 3: Add Left Bar Button Item to restart game

I also went ahead to add some AlertController for slightly better UX. I
reminded myself that handlers in UIAlertAction does not require the
`selector` keyword, and that you can force unwrap the UIAlertAction in
the parameter to assign default of nil

* Bonus: Fix bugs pertaining to word cases

This is the laziest way to fix the problem, because it will force
whatever input to become lowercase.

An alternative way to do this is to really put the answer the way user
inputs it, but in the `isOriginal` method, you need to create a copy of
the array with all the words lowercased. In terms of time complexity,
this is not really efficient. I suppose this is a trade-off that can be
made.