Add button to clear the shopping list
To create the handler for the confirm UIAlertAction, I tried using
closures. Initially I just used a single parameter `_`, and tried
accessing `shoppingList` as well as `tableView` directly in the closure.
However it gave me an error.

The error is along the lines of: "Reference to property in closure
requires explicit use of 'self' to make capture semantics explicit."
This was when I opted to use the capture list with a weak reference to
self. But again, I used weak reference because this is how it was done
all along.

Additional note: I wonder how different it is to call the `removeAll()`
method and assigning a new empty array to the shoppingList array.