Prepare ActivityViewController for sharing
Here, I discovered the `rightBarButtonItems` property of
UINavigationItem. This allows multiple BarButtonItems instead of a
single one. I initially wanted to use a toolbar for the share, but this simplifies things a lot!

A small challenge I faced is in the handling of the
popoverPresentationController for iPad. I usually set it to the
rightBarButtonItem, but that will cause it to pop-up to the right-most
item. The share button is the second right-most, so I need to index the
array.

Currently, the index is hardcoded as I know exactly where it is.
However, perhaps a better way would be to have some properties that
dictates the position of such BarButtonItems in the array. That way, if
I ever feel like swapping the position of the share button, I can simply
change the index in the array.

An alternative I am thinking is to find the indexOf the button that I
want in the Array. I think this is tricky to do because it requires some
understanding of the equality of the UI BarButtonItem