Complete Milestone 2: Projects 4 - 6 (#7)
* Add Milestone2 starter files

* Initialise TableViewController for the Shopping List

This one is pretty standard! I was able to do this without any
reference, but I did forget to initialise the entry point of the
ViewController.

* Add BarButton to add new items to the list

This one required significantly more reference to Project 5. I was
reminded of the `addTextField()` method, and perhaps I require some
review of why a weak reference is needed in this context.

* Add button to clear the shopping list

To create the handler for the confirm UIAlertAction, I tried using
closures. Initially I just used a single parameter `_`, and tried
accessing `shoppingList` as well as `tableView` directly in the closure.
However it gave me an error.

The error is along the lines of: "Reference to property in closure
requires explicit use of 'self' to make capture semantics explicit."
This was when I opted to use the capture list with a weak reference to
self. But again, I used weak reference because this is how it was done
all along.

Additional note: I wonder how different it is to call the `removeAll()`
method and assigning a new empty array to the shoppingList array.

* Prepare ActivityViewController for sharing

Here, I discovered the `rightBarButtonItems` property of
UINavigationItem. This allows multiple BarButtonItems instead of a
single one. I initially wanted to use a toolbar for the share, but this simplifies things a lot!

A small challenge I faced is in the handling of the
popoverPresentationController for iPad. I usually set it to the
rightBarButtonItem, but that will cause it to pop-up to the right-most
item. The share button is the second right-most, so I need to index the
array.

Currently, the index is hardcoded as I know exactly where it is.
However, perhaps a better way would be to have some properties that
dictates the position of such BarButtonItems in the array. That way, if
I ever feel like swapping the position of the share button, I can simply
change the index in the array.

An alternative I am thinking is to find the indexOf the button that I
want in the Array. I think this is tricky to do because it requires some
understanding of the equality of the UI BarButtonItem

* Join strings in shoppingList that can be used for sharing

Looks like you can't share arrays in activityItems!

* Add cancel button when trying to add for slightly better UX

* Prevent empty strings from being added to the list