Complete Project 7 (#8)
* Add Project7 starter files

* Add basic Project7

Note that the changes to AppDelegate should be made to SceneDelegate
instead

* Challenge 1: Add Credits button with UIBarButtonItem

Pretty standard!

* Challenge 2: Prepare additional filtered array and states

I expect it to be some Boolean checks. Once I click on filter, it will
switch state to filtered state and only show the contents of the filteredPetitions

* Challenge 2: Add method to fill in the filteredPetitions

This requires changing the filtered flag as well as reloading the
tableView

* Challenge 2: Add method to clear filter

I got a bit lazy in resetting the button to the original filter button,
so I reused the old method of trying to configure the navbar. I probably
should abstract it a bit so that it can be easily reused.

* Challenge 2: Refactor methods to change the LeftBarButtonItem

* Challenge 2: Include filter keyword in the title of NavBar

However, this won't work for very long filter keywords. Will probably
need another way to show this to the user.

* Update README for Project7